Bonuses added!
Updated the gitignore file to ignore .diff files (personal preference to long commits
so I don't forget things)

Updated the CollisionModel to be easier for classes to use. More self contained

Added a bonus for testing in the game screen. Added fallspeed initialization.

Remembered to set the collision flags for player.

Changed the CollisionModel's collision flag to be an unsigned int so things work.