And Clouds were then born!
Clouds sparked the addition of the "mRenderables" vector. Because clouds should *not* be
even considered for collision, and adding a shitload of them caused the main game loop
to skyrocket.... because of the n*n collision :< So I have to refactor some of that
next. Bummer.

A cloud is basically a scenery. Nothing more.